rss-story-inbox/services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-dateutil==2.9.0.post0

feedparser==6.0.11
trafilatura==1.12.2
rapidfuzz==3.9.6

openai==1.40.6
httpx==0.27.0
